shroomme/requirements.txt
Django==1.8
Pillow==2.9.0
PyJWT==1.4.0
django-crispy-forms==1.5.2
django-registration-redux==1.2
mongoengine==0.8.0
oauthlib==1.0.3
pymongo==3.0.3
python-openid==2.2.5
python-social-auth==0.2.12
requests==2.7.0
requests-oauthlib==0.5.0
six==1.9.0
wheel==0.24.0
